login, signup, scoreboard, packages: done

+++ src/main/java/game_server/db/DbConnection.java
@@ -3,6 +3,7 @@ package game_server.db;
 import game.cards.Card;
 import game.cards.MonsterCard;
 import game.cards.SpellCard;
+import game.decks.CardDeck;
 import game.decks.CardStack;
 import game.enums.Element;
 import game.enums.MonsterType;
@@ -58,7 +59,6 @@ public class DbConnection {
             String fullPath = "jdbc:postgresql://" + this.dbUrl + ":"+ this.port +"/" + this.dbName;
             Connection c = DriverManager.getConnection( fullPath, this.username, this.password);
             c.setAutoCommit(false);
-            System.out.println("--> Opened database successfully");
             return c;
         } catch ( Exception e ) {
             System.err.println( e.getClass().getName()+": "+ e.getMessage() );
@@ -67,52 +67,30 @@ public class DbConnection {
         }
     }
 
-    public boolean insertUser(User player) {
-        c = connect();
+    public boolean insertUser(User player) throws SQLException {
         int rows = 0;
-        try {
-            String query = ("INSERT INTO PUBLIC.USER " +
-                            "(username, password, bio, token, image, coins, elo) " +
-                            "VALUES (?, ?, ?, ?, ?, ?, ?);");
-            PreparedStatement stmt = c.prepareStatement(query);
-            stmt.setString(1, player.getUsername());
-            stmt.setString(2, player.getPassword());
-            stmt.setString(3, player.getBio());
-            stmt.setString(4, player.getToken());
-            stmt.setString(5, player.getImage());
-            stmt.setInt(6, player.getCoins());
-            stmt.setInt(7, player.getElo());
-            rows = stmt.executeUpdate();
 
-            if (rows > 0) c.commit();
-            closeConnection(stmt);
-            c = null;
+        this.c = connect();
+        String query = ("INSERT INTO PUBLIC.USER " +
+                        "(username, password, bio, token, image, coins, elo, admin) " +
+                        "VALUES (?, ?, ?, ?, ?, ?, ?, ?);");
+        PreparedStatement stmt = this.c.prepareStatement(query);
+        stmt.setString(1, player.getUsername());
+        stmt.setString(2, player.getPassword());
+        stmt.setString(3, player.getBio());
+        stmt.setString(4, player.getToken());
+        stmt.setString(5, player.getImage());
+        stmt.setInt(6, player.getCoins());
+        stmt.setInt(7, player.getElo());
+        stmt.setBoolean(8, player.isAdmin());
 
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-        }
+        rows = stmt.executeUpdate();
+
+        if (rows > 0) this.c.commit();
+        closeConnection(stmt);
+        this.c = null;
+
         return rows > 0;
-    }
-
-    public boolean isLogged(String token) {
-        try {
-            c = connect();
-            String query = ( "SELECT * FROM PUBLIC.SESSION WHERE token = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
-            stmt.setString(1, token);
-            ResultSet rs = stmt.executeQuery();
-            try {
-                if(rs.next()) {
-                    return true;
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-            closeConnection(rs, stmt);
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-        }
-        return false;
     }
 
     public User getLoggedUser(String token) {
@@ -128,31 +106,36 @@ public class DbConnection {
         return user;
     }
 
-    public boolean insertCard(Card card, Boolean toDeck, User user, String cid) {
-        c = connect();
+    public boolean insertCard(Card card, Boolean toDeck, User user, String cid, int packageId) {
+        this.c = connect();
         int rows = 0;
+
         try {
             String query = (
             "INSERT INTO public.card" +
             " (\"cardId\", name, \"monsterType\", element, damage, deck, owner, \"packageId\")" +
             " VALUES (?, ?, ?, ?, ?, ?, ?, ?);");
 
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, cid);
             stmt.setString(2, card.getName());
-            stmt.setString(3, card.getType().getName());
+
+            String type = (card.getType() == null) ? null : card.getType().getName();
+            stmt.setString(3, type);
             stmt.setString(4,  card.getCardElement().getElementName());
             stmt.setFloat(5, card.getDamage());
             stmt.setBoolean(6, toDeck);
-            stmt.setString(7, user.getUsername());
-            stmt.setInt(8, getMaxPackageId() + 1);
+
+            String owner = (user.isAdmin()) ? null : user.getUsername();
+            stmt.setString(7, owner);
+            stmt.setInt(8, packageId);
             rows = stmt.executeUpdate();
 
-            if (rows > 0) c.commit();
+            if (rows > 0) this.c.commit();
             closeConnection(stmt);
 
         } catch (SQLException throwables) {
-            throwables.printStackTrace();
+            System.out.println("This card were already loaded to DB");
         }
         return rows > 0;
     }
@@ -210,8 +193,9 @@ public class DbConnection {
 
     public int getMaxPackageId() {
         try {
+            this.c = connect();
             String query = ("SELECT max(\"packageId\") as maxId FROM PUBLIC.CARD;");
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             ResultSet rs = stmt.executeQuery();
             try {
                 if(rs.next()) {
@@ -219,43 +203,64 @@ public class DbConnection {
                 }
             } catch (Exception e) {
                 e.printStackTrace();
+                return 0;
             }
+            closeConnection(rs,stmt);
         } catch (RuntimeException | SQLException e) {
-            System.out.println("Exception");
+            System.out.println("Get max packageId exception");
+            return 0;
         }
         return 0;
     }
 
     public boolean addSession(String token){
-        c = connect();
         int rows = 0;
         try {
-            String query = ("INSERT INTO PUBLIC.SESSION (token) VALUES (?);");
-            PreparedStatement stmt = c.prepareStatement(query);
-            stmt.setString(1, token);
-            rows = stmt.executeUpdate();
+            if(!isLogged(token)) {
+                this.c = connect();
+                String query = ("INSERT INTO PUBLIC.SESSION (token) VALUES (?);");
+                PreparedStatement stmt = this.c.prepareStatement(query);
+                stmt.setString(1, token);
+                rows = stmt.executeUpdate();
+                if (rows > 0) this.c.commit();
 
-            if (rows > 0) c.commit();
-            closeConnection(stmt);
-
+                closeConnection(stmt);
+                return true;
+            }
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
-        return rows > 0;
+        return false;
+    }
+
+    public boolean isLogged(String token) throws SQLException { //getSession
+        this.c = connect();
+        String query = ("SELECT * FROM PUBLIC.SESSION WHERE token = ?;");
+        PreparedStatement stmt = this.c.prepareStatement(query);
+        stmt.setString(1, token);
+        ResultSet rs = stmt.executeQuery();
+
+        if (rs.next()) {
+            closeConnection(rs,stmt);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public boolean deleteSession(String token) {
-        c = connect();
         int rows = 0;
         try {
-            String query = ("DELETE FROM public.session WHERE token LIKE ?;");
-            PreparedStatement stmt = c.prepareStatement(query);
-            stmt.setString(1, token);
-            rows = stmt.executeUpdate();
+            if(isLogged(token)) {
+                this.c = connect();
+                String query = ("DELETE FROM public.session WHERE token LIKE ?;");
+                PreparedStatement stmt = this.c.prepareStatement(query);
+                stmt.setString(1, token);
+                rows = stmt.executeUpdate();
 
-            if (rows > 0) c.commit();
-            closeConnection(stmt);
-
+                if (rows > 0) this.c.commit();
+                closeConnection(stmt);
+            }
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
@@ -265,15 +270,15 @@ public class DbConnection {
     public User getUser(String username, String pass) {
         User user = null;
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "SELECT * FROM PUBLIC.USER WHERE username = ? AND password = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, username);
             stmt.setString(2, pass);
             ResultSet rs = stmt.executeQuery();
             try {
                 if(rs.next()) {
-                    user = buildUser(rs);
+                    user = buildUser(rs, true);
                 }
             } catch (Exception e) {
                 e.printStackTrace();
@@ -288,14 +293,14 @@ public class DbConnection {
     public User getUser(String username) {
         User user = null;
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "SELECT * FROM PUBLIC.USER WHERE username = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, username);
             ResultSet rs = stmt.executeQuery();
             try {
                 if(rs.next()) {
-                    user = buildUser(rs);
+                    user = buildUser(rs, false);
                 }
             } catch (Exception e) {
                 e.printStackTrace();
@@ -307,13 +312,36 @@ public class DbConnection {
         return user;
     }
 
-    public Card getCardById(String cid) throws SQLException {
+    public List<User> getAllUsers() {
+        List<User> userList = new ArrayList<>();
+        try {
+            this.c = connect();
+            String query = ( "SELECT * FROM PUBLIC.USER;" );
+            PreparedStatement stmt = this.c.prepareStatement(query);
+            ResultSet rs = stmt.executeQuery();
+            try {
+                if (rs.next()) {
+                    do {
+                        userList.add(buildUser(rs, false));
+                    } while (rs.next());
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            closeConnection(rs, stmt);
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
+        }
+        return userList;
+    }
+
+    public Card getCardById(String cid) {
         Card card = null;
 
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "SELECT * FROM PUBLIC.CARD WHERE \"cardId\" = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, cid);
             ResultSet rs = stmt.executeQuery();
             if(rs == null || !rs.next()) {
@@ -321,7 +349,7 @@ public class DbConnection {
                 return null;
             }
 
-            card = initCard(rs);
+            card = buildCard(rs);
             closeConnection(rs, stmt);
         } catch (SQLException throwables) {
             throwables.printStackTrace();
@@ -330,19 +358,19 @@ public class DbConnection {
         return card;
     }
 
-    public List<Card> getCardByOwner(String username) throws SQLException {
+    public List<Card> getCardByOwner(String username) {
         List<Card> cardList = new ArrayList<>();
         Card temp;
 
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "SELECT * FROM PUBLIC.CARD WHERE \"owner\" = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, username);
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
                 do {
-                    temp = initCard(rs);
+                    temp = buildCard(rs);
                     cardList.add(temp);
                 } while (rs.next());
             }
@@ -355,16 +383,60 @@ public class DbConnection {
         return cardList;
     }
 
-    private Card initCard(ResultSet rs) throws SQLException {
+    public List<Card> getPackageCards(int packageId) {
+        List<Card> cardList = new ArrayList<>();
+        Card temp;
+
+        try {
+            this.c = connect();
+            String query = ( "SELECT * FROM PUBLIC.CARD WHERE \"packageId\" = ?;" );
+            PreparedStatement stmt = this.c.prepareStatement(query);
+            stmt.setInt(1, packageId);
+            ResultSet rs = stmt.executeQuery();
+            if (rs.next()) {
+                do {
+                    temp = buildCard(rs);
+                    cardList.add(temp);
+                } while (rs.next());
+            }
+
+            closeConnection(rs, stmt);
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
+        }
+
+        return cardList;
+    }
+
+    public void addOwnerToPackage(String username, int packageId) {
+        try {
+            this.c = connect();
+            String query = (
+                "UPDATE public.card SET owner = ?, \"packageId\" = null  WHERE \"packageId\" = ?;" );
+            PreparedStatement stmt = c.prepareStatement(query);
+            stmt.setString(1, username);
+            stmt.setInt(2, packageId);
+            int rows = stmt.executeUpdate();
+            if (rows > 0) c.commit();
+
+            closeConnection(stmt);
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
+        }
+    }
+
+    private Card buildCard(ResultSet rs) throws SQLException {
         Card card;
 
         if ( rs.getString("monsterType") == null) {
             card = new SpellCard(
+                    rs.getString("cardId"),
                     Element.find(rs.getString("element")),
                     rs.getString("name"),
                     rs.getFloat("damage"));
         } else {
             card = new MonsterCard(
+                    rs.getString("cardId"),
                     Objects.requireNonNull(MonsterType.find(rs.getString("monsterType"))),
                     Objects.requireNonNull(Element.find(rs.getString("element"))),
                     rs.getString("name"),
@@ -373,15 +445,23 @@ public class DbConnection {
         return card;
     }
 
-    public User buildUser(ResultSet rs) throws SQLException {
+    public User buildUser(ResultSet rs, boolean withPass) throws SQLException {
+        String pass;
+        if(withPass) {
+            pass = rs.getString("password");
+        } else {
+            pass = "";
+        }
         User user = User.builder()
                 .username( rs.getString("username") )
-                .password( rs.getString("password") )
+                .password( pass )
                 .token( rs.getString("token") )
                 .bio( rs.getString("bio") )
                 .coins( rs.getInt("coins") )
                 .elo( rs.getInt("elo") )
                 .stack(new CardStack())
+                .deck(new CardDeck())
+                .isAdmin( rs.getBoolean("admin") )
                 .build();
         List<Card> list = getCardByOwner(user.getUsername());
         user.getStack().addListToStack(list);
@@ -413,15 +493,15 @@ public class DbConnection {
 
     public boolean deleteUser(User user) {
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "DELETE FROM PUBLIC.USER WHERE username = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, user.getUsername());
             int rows = stmt.executeUpdate();
 
-            if (rows > 0) c.commit();
+            if (rows > 0) this.c.commit();
             closeConnection(stmt);
-            c = null;
+            this.c = null;
 
             return true;
         } catch (SQLException throwables) {
@@ -432,13 +512,13 @@ public class DbConnection {
 
     public boolean deleteCard(String cid) {
         try {
-            c = connect();
+            this.c = connect();
             String query = ( "DELETE FROM PUBLIC.CARD WHERE \"cardId\" = ?;" );
-            PreparedStatement stmt = c.prepareStatement(query);
+            PreparedStatement stmt = this.c.prepareStatement(query);
             stmt.setString(1, cid);
             int rows = stmt.executeUpdate();
 
-            if (rows > 0) c.commit();
+            if (rows > 0) this.c.commit();
             closeConnection(stmt);
 
             return true;
@@ -451,12 +531,12 @@ public class DbConnection {
     public void closeConnection(ResultSet rs, PreparedStatement stmt) throws SQLException {
         rs.close();
         stmt.close();
-        c.close();
+        this.c.close();
     }
 
     public void closeConnection(PreparedStatement stmt) throws SQLException {
         stmt.close();
-        c.close();
+        this.c.close();
     }
 
 }
